Re-designed pixel art selection algorithm to be aribtrary
Original pixel selection required a rect area to be defined, now all
operations can be performed on arbitrary arrays of points.

Changed the default splat icon to be a box, (it looks more object-ey)

extracted the continuous pixel algorithm from the fill tool, and added
it as a canvas function, with definable region.

Cleaned out comments, ect.